+++ src/Symfony/Component/Messenger/MessageBusInterface.php
@@ -19,11 +19,9 @@ interface MessageBusInterface
     /**
      * Dispatches the given message.
      *
-     * The bus can return a value coming from handlers, but is not required to do so.
-     *
      * @param object|Envelope $message The message or the message pre-wrapped in an envelope
-     *
-     * @return mixed
+     * @param string|null     $name    The name to use as dispatching key; when not provided,
+     *                                 this name is derived from the type of the message
      */
-    public function dispatch($message);
+    public function dispatch($message, string $name = null): void;
 }