+++ src/Symfony/Component/Messenger/Tests/Handler/ChainHandlerTest.php
@@ -12,33 +12,30 @@
 namespace Symfony\Component\Messenger\Tests\Handler;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Handler\ChainHandler;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 
 class ChainHandlerTest extends TestCase
 {
-    public function testItCallsTheHandlersAndReturnsAllResults()
+    public function testItCallsTheHandlers()
     {
-        $message = new DummyMessage('Hey');
+        $message = new Envelope(new DummyMessage('Hey'));
 
         $handler1 = $this->createPartialMock(\stdClass::class, array('__invoke'));
         $handler1
             ->expects($this->once())
             ->method('__invoke')
             ->with($message)
-            ->willReturn('Hello')
         ;
         $handler2 = $this->createPartialMock(\stdClass::class, array('__invoke'));
         $handler2
             ->expects($this->once())
             ->method('__invoke')
             ->with($message)
-            ->willReturn('World')
         ;
 
-        $results = (new ChainHandler(array($handler1, $handler2)))($message);
-
-        $this->assertSame(array('Hello', 'World'), $results);
+        (new ChainHandler(array($handler1, $handler2)))($message);
     }
 
     /**